% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Fishing}
\alias{Fishing}
\title{Choice of Fishing Mode}
\format{
A data frame containing :

- mode: recreation mode choice, one of : beach, pier, boat and charter,
- price.beach: price for beach mode
- price.pier: price for pier mode,
- price.boat: price for private boat mode,
- price.charter: price for charter boat mode,
- catch.beach: catch rate for beach mode,
- catch.pier: catch rate for pier mode,
- catch.boat: catch rate for private boat mode,
- catch.charter: catch rate for charter boat mode,
- income: monthly income,
}
\source{
Cameron A, Trivedi P (2005). Microeconometrics. Cambridge University Press.
https://EconPapers.repec.org/RePEc:cup:cbooks:9780521848053.
}
\description{
A sample of 1182 individuals in the United-States for the choice of
4 alternative fishing modes.
}
\references{
Herriges JA, Kling CL (1999). “Nonlinear Income Effects in Random Utility
Models.” The Review of Economics and Statistics, 81(1), 62-72.
doi: 10.1162/003465399767923827
}
\seealso{
\code{\link{retinopathy}}, \code{\link{vaso}}
}
\keyword{dataset}
