% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.brant}
\alias{print.brant}
\title{Print method for an object of class brant}
\usage{
\method{print}{brant}(x, ...)
}
\arguments{
\item{x}{An object of class \code{brant}.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints out a summary table of brant-test for an object of class brant.
}
\seealso{
\code{\link{brant.test}}
}
