% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.erroR}
\alias{print.erroR}
\title{Print method for an object of class erroR}
\usage{
\method{print}{erroR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{erroR}.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints out a vector of name and calculated error value of fitted model.
}
\seealso{
\code{\link{erroR}}
}
