% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.hosmerlem}
\alias{print.hosmerlem}
\title{Print method for an object of class hosmerlem}
\usage{
\method{print}{hosmerlem}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hosmerlem}.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints out a summary table of the goodness-of-fit test for binary,
multinomial and ordinal models.
class LRI.
}
\seealso{
\code{\link{hosmerlem}}
}
