% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.lipsitz}
\alias{print.lipsitz}
\title{Print method for an object of class lipsitz}
\usage{
\method{print}{lipsitz}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lipsitz}.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints out a summary table of the goodness-of-fit test for an object of
class lipsitz.
}
\seealso{
\code{\link{lipsitz}}
}
