% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.LRT}
\alias{print.LRT}
\title{Print method for an object of class LRT}
\usage{
\method{print}{LRT}(x, ...)
}
\arguments{
\item{x}{An object of class \code{LRI}.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints out a summary table of the likelihood ratio test for an object of
class LRI.
}
\seealso{
\code{\link{LR.test}}
}
