% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.pulkroben}
\alias{print.pulkroben}
\title{Print method for an object of class pulkroben}
\usage{
\method{print}{pulkroben}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pulkroben}.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints out a summary table of the goodness-of-fit test for an object of
class pulkroben.
}
\seealso{
\code{\link{pulkroben}}
}
