% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.Rsquared}
\alias{print.Rsquared}
\title{Print method for an object of class r-squared}
\usage{
\method{print}{Rsquared}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rsquared}.}

\item{...}{additional arguments.}
}
\value{
No return value
}
\description{
Prints out a vector of name and value of an R2 of a fitted model.
}
\seealso{
\code{\link{Rsquared}}
}
