% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{retinopathy}
\alias{retinopathy}
\title{Retinopathy}
\format{
A data frame with 613 observations on the following 5 variables:
- RET: RET=1: no retinopathy, RET=2 nonproliferative retinopathy,
RET=3 advanced retinopathy or blind
- SM: SM=1: smoker, SM=0: non-smoker
- DIAB: diabetes duration in years
- GH: glycosylated hemoglobin measured in percent
- BP: diastolic blood pressure in mmHg
}
\description{
The \code{retinopathy} data contains information on persons with retinopathy.
}
\references{
Bender and  Grouven (1998), Using binary logistic regression models for
ordinal data with non-proportional odds, \emph{J. Clin. Epidemiol.},
51, 809-816.
}
\seealso{
\code{\link{vaso}}, \code{\link{Fishing}}
}
\keyword{dataset}
