% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquared.R
\name{Rsquared}
\alias{Rsquared}
\title{R-squared for Categorical Response Models}
\usage{
Rsquared(model, measure)
}
\arguments{
\item{model}{single model object for which R2 is determined.}

\item{measure}{selects any of the different measures available.}
}
\value{
\item{measure}{the name of the R-squared calculated.}

\item{R2}{realized value of the computed R2.}

\item{adj}{adjusted R2, only available when McFadden's R2 is computed.}
}
\description{
computes the summary measures of predictive strength of
several categorical outcome model. R-squared measures are computed based
on suitability to the model under consideration.
}
\details{
\code{Rsquared} provides different R2 indices for both binary and
multi-categorical response models. Supported classes include: \code{glm},
\code{vglm}, \code{clm}, \code{polr}, \code{multinom}, \code{mlogit},
\code{serp}. Thus, mainly model families with binary and multi-categorical
response are supported.
}
\examples{
require(serp)

pom <- serp(ordered(RET) ~ DIAB + GH + BP, link="logit",
            slope = "parallel", reverse = TRUE, data = retinopathy)
Rsquared(pom, measure = "mcfadden")
Rsquared(pom, measure = "ugbagerth")

}
\references{
Long, J.S. (1997). \emph{Regression Models for Categorical and Limited
    Dependent Variables}. California: Sage Publications.
}
\seealso{
\code{\link{erroR}}
}
