% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vaso}
\alias{vaso}
\title{Vasoconstriction and Breathing}
\format{
A data frame with 39 observations on the following 3 variables:
- vol: amount of air
- rate: rate of breathing
- vaso: condition of vasculature: no vasoconstriction = 1, vasoconstriction = 2
}
\source{
Data Archive Department of Statistics, LMU Munich.
}
\description{
A binary data on the neural constriction of vasculature. Three test persons
inhaled a certain amount of air with different rates. In some cases a
vasoconstriction occurred at their skin. The objective of the study was to
indicate a correlation between breathing and vasoconstriction.
The test persons repeated the test 9, 8, 22 times, making a total of
39 observations.
}
\references{
Finney, D. J. (1971) \emph{Probit Analysis}. 3rd edition. Cambridge
University Press.

Pregibon, D. (1982) Resistant fits for some commonly used logistic models.
Appl. Stat. \bold{29}, 15--24.

Hastie, T. J. and Tibshirani, R. J. (1990) \emph{Generalized Additve Models}.
Chapman and Hall.
}
\seealso{
\code{\link{retinopathy}}, \code{\link{Fishing}}
}
\keyword{dataset}
